refactor(ingest/snowflake): move schema extraction to dedicated file
