+++ metadata-ingestion/src/datahub/ingestion/source/snowflake/snowflake_schema.py
@@ -3,10 +3,11 @@ from collections import defaultdict
 from dataclasses import dataclass, field
 from datetime import datetime
 from functools import lru_cache
-from typing import Dict, List, Optional
+from typing import Callable, Dict, List, Optional
 
 from snowflake.connector import SnowflakeConnection
 
+from datahub.ingestion.api.report import SupportsAsObj
 from datahub.ingestion.source.snowflake.constants import SnowflakeObjectDomain
 from datahub.ingestion.source.snowflake.snowflake_query import SnowflakeQuery
 from datahub.ingestion.source.snowflake.snowflake_utils import SnowflakeQueryMixin
@@ -176,7 +177,7 @@ class _SnowflakeTagCache:
         )
 
 
-class SnowflakeDataDictionary(SnowflakeQueryMixin):
+class SnowflakeDataDictionary(SnowflakeQueryMixin, SupportsAsObj):
     def __init__(self) -> None:
         self.logger = logger
         self.connection: Optional[SnowflakeConnection] = None
@@ -188,6 +189,26 @@ class SnowflakeDataDictionary(SnowflakeQueryMixin):
         # Connection is already present by the time this is called
         assert self.connection is not None
         return self.connection
+
+    def as_obj(self) -> Dict[str, Dict[str, int]]:
+        # TODO: Move this into a proper report type that gets computed.
+
+        # Reports how many times we reset in-memory `functools.lru_cache` caches of data,
+        # which occurs when we occur a different database / schema.
+        # Should not be more than the number of databases / schemas scanned.
+        # Maps (function name) -> (stat_name) -> (stat_value)
+        lru_cache_functions: List[Callable] = [
+            self.get_tables_for_database,
+            self.get_views_for_database,
+            self.get_columns_for_schema,
+            self.get_pk_constraints_for_schema,
+            self.get_fk_constraints_for_schema,
+        ]
+
+        report = {}
+        for func in lru_cache_functions:
+            report[func.__name__] = func.cache_info()._asdict()  # type: ignore
+        return report
 
     def show_databases(self) -> List[SnowflakeDatabase]:
         databases: List[SnowflakeDatabase] = []
