+++ metadata-ingestion/src/datahub/ingestion/source/snowflake/snowflake_utils.py
@@ -7,6 +7,7 @@ from typing_extensions import Protocol
 
 from datahub.configuration.common import MetaError
 from datahub.configuration.pattern_utils import is_schema_allowed
+from datahub.emitter.mce_builder import make_dataset_urn_with_platform_instance
 from datahub.ingestion.source.snowflake.constants import (
     GENERIC_PERMISSION_ERROR_KEY,
     SNOWFLAKE_REGION_CLOUD_REGION_MAPPING,
@@ -48,6 +49,8 @@ class SnowflakeQueryMixin:
 
 
 class SnowflakeCommonProtocol(SnowflakeLoggingProtocol, Protocol):
+    platform: str = "snowflake"
+
     config: SnowflakeV2Config
     report: SnowflakeV2Report
 
@@ -177,6 +180,14 @@ class SnowflakeCommonMixin:
         if self.config.convert_urns_to_lowercase:
             return identifier.lower()
         return identifier
+
+    def gen_dataset_urn(self: SnowflakeCommonProtocol, dataset_identifier: str) -> str:
+        return make_dataset_urn_with_platform_instance(
+            platform=self.platform,
+            name=dataset_identifier,
+            platform_instance=self.config.platform_instance,
+            env=self.config.env,
+        )
 
     @staticmethod
     def get_quoted_identifier_for_database(db_name):
